% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_to_json.R
\name{ds_to_json}
\alias{ds_to_json}
\title{ds_to_json}
\usage{
ds_to_json(x)
}
\arguments{
\item{x}{A named list to be converted to json data structure}
}
\value{
a string in json format
}
\description{
To convert a named list into a json data structure
}
